requirements.txt
enum34==1.0.4
Flask==0.10.1
Flask-Assets==0.10
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.8
Markdown==2.4
MarkupSafe==0.23
pathlib==1.0.1
PyGithub==1.25.2
pyScss==1.3.4
PyYAML==3.11
six==1.9.0
webassets==0.10.1
Werkzeug==0.9.4
wheel==0.24.0
